requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
Django==4.2.2
django-redis==5.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.0
